refactor: interactions
