+++ Cargo.toml
@@ -11,9 +11,9 @@ serde_json = "1.0.81"
 serde = { version = "1.0.137", features = ["derive"] }
 regex = "1.5.6"
 urlencoding = "2.1.0"
-sqlx = { version = "0.5.13", features = ["runtime-tokio-rustls", "sqlite", "offline"] }
+sqlx = { version = "0.6.2", features = ["runtime-tokio-rustls", "sqlite", "offline"] }
 anyhow = "1.0.57"
-meilisearch-sdk = "0.17.0"
+meilisearch-sdk = "0.20.1"
 words-count = "0.1.3"
 html-escape = "0.2.11"
 piston_rs = "0.4.2"
